做完 bomb 和 attack 两个实验

+++ malloclab-handout/mdriver.c
@@ -31,7 +31,7 @@
 #define LINENUM(i) (i+5) /* cnvt trace request nums to linenums (origin 1) */
 
 /* Returns true if p is ALIGNMENT-byte aligned */
-#define IS_ALIGNED(p)  ((((unsigned int)(p)) % ALIGNMENT) == 0)
+#define IS_ALIGNED(p)  ((((unsigned long)(p)) % ALIGNMENT) == 0)
 
 /****************************** 
  * The key compound data types
